# frontend/src/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Squant",
    "description": "Personal Quantitative Trading System for Cryptocurrency",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/market/exchange/{exchange_id}": {
      "put": {
        "tags": [
          "Market"
        ],
        "summary": "Set Exchange",
        "description": "Switch data source exchange.\n\nFor single-user system, this dynamically switches the exchange for\nboth REST API and WebSocket data sources.",
        "operationId": "set_exchange_api_v1_market_exchange__exchange_id__put",
        "parameters": [
          {
            "name": "exchange_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Exchange ID (okx, binance, bybit)",
              "title": "Exchange Id"
            },
            "description": "Exchange ID (okx, binance, bybit)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exchange-accounts": {
      "post": {
        "tags": [
          "Exchange Accounts"
        ],
        "summary": "Create Exchange Account",
        "description": "Create a new exchange account configuration.\n\nStore API credentials securely with AES-256-GCM encryption.\n\nArgs:\n    request: Account creation request with credentials.\n    session: Database session.\n\nReturns:\n    Created account (credentials not returned).\n\nRaises:\n    HTTPException: 400 if OKX without passphrase, 409 if name exists.",
        "operationId": "create_exchange_account_api_v1_exchange_accounts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExchangeAccountRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ExchangeAccountResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Exchange Accounts"
        ],
        "summary": "List Exchange Accounts",
        "description": "List all exchange account configurations.\n\nArgs:\n    exchange: Optional exchange filter.\n    session: Database session.\n\nReturns:\n    List of accounts (credentials not returned).",
        "operationId": "list_exchange_accounts_api_v1_exchange_accounts_get",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "okx",
                    "binance"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exchange",
              "title": "Exchange"
            },
            "description": "Filter by exchange"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ExchangeAccountListItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exchange-accounts/{account_id}": {
      "get": {
        "tags": [
          "Exchange Accounts"
        ],
        "summary": "Get Exchange Account",
        "description": "Get exchange account details by ID.\n\nArgs:\n    account_id: Account UUID.\n    session: Database session.\n\nReturns:\n    Account details (credentials not returned).\n\nRaises:\n    HTTPException: 404 if not found.",
        "operationId": "get_exchange_account_api_v1_exchange_accounts__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ExchangeAccountResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Exchange Accounts"
        ],
        "summary": "Update Exchange Account",
        "description": "Update an exchange account configuration.\n\nArgs:\n    account_id: Account UUID.\n    request: Update request.\n    session: Database session.\n\nReturns:\n    Updated account (credentials not returned).\n\nRaises:\n    HTTPException: 404 if not found, 409 if name exists.",
        "operationId": "update_exchange_account_api_v1_exchange_accounts__account_id__put",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExchangeAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ExchangeAccountResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exchange Accounts"
        ],
        "summary": "Delete Exchange Account",
        "description": "Delete an exchange account configuration.\n\nArgs:\n    account_id: Account UUID.\n    session: Database session.\n\nReturns:\n    Success response.\n\nRaises:\n    HTTPException: 404 if not found, 409 if account is in use.",
        "operationId": "delete_exchange_account_api_v1_exchange_accounts__account_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exchange-accounts/{account_id}/test": {
      "post": {
        "tags": [
          "Exchange Accounts"
        ],
        "summary": "Test Exchange Connection",
        "description": "Test connection to exchange using stored credentials.\n\nAttempts to fetch account balance to verify API credentials are valid.\n\nArgs:\n    account_id: Account UUID.\n    session: Database session.\n\nReturns:\n    Connection test result with success status and balance count.\n\nRaises:\n    HTTPException: 404 if account not found.",
        "operationId": "test_exchange_connection_api_v1_exchange_accounts__account_id__test_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConnectionTestResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_ConnectionTestResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ConnectionTestResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[ConnectionTestResponse]"
      },
...
      "ApiResponse_ExchangeAccountResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ExchangeAccountResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[ExchangeAccountResponse]"
      },
...
      "ApiResponse_NoneType_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[NoneType]"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[dict]"
      },
...
      "ApiResponse_list_ExchangeAccountListItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ExchangeAccountListItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[ExchangeAccountListItem]]"
      },
...
      "ConnectionTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the connection test succeeded"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Error message if failed"
          },
          "balance_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance Count",
            "description": "Number of currencies with balance (if successful)"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ConnectionTestResponse",
        "description": "Response for exchange connection test."
      },
...
      "CreateExchangeAccountRequest": {
        "properties": {
          "exchange": {
            "type": "string",
            "enum": [
              "okx",
              "binance",
              "bybit"
            ],
            "title": "Exchange",
            "description": "Exchange identifier (okx, binance, or bybit)"
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "description": "Unique account name"
          },
          "api_key": {
            "type": "string",
            "format": "password",
            "title": "Api Key",
            "description": "Exchange API key",
            "writeOnly": true
          },
          "api_secret": {
            "type": "string",
            "format": "password",
            "title": "Api Secret",
            "description": "Exchange API secret",
            "writeOnly": true
          },
          "passphrase": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Passphrase",
            "description": "API passphrase (required for OKX)"
          },
          "testnet": {
            "type": "boolean",
            "title": "Testnet",
            "description": "Whether to use testnet/sandbox mode",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "exchange",
          "name",
          "api_key",
          "api_secret"
        ],
        "title": "CreateExchangeAccountRequest",
        "description": "Request to create a new exchange account configuration."
      },
...
      "ExchangeAccountListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "testnet": {
            "type": "boolean",
            "title": "Testnet"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exchange",
          "name",
          "testnet",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "ExchangeAccountListItem",
        "description": "Exchange account item for list response."
      },
      "ExchangeAccountResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "testnet": {
            "type": "boolean",
            "title": "Testnet"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exchange",
          "name",
          "testnet",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "ExchangeAccountResponse",
        "description": "Exchange account details response.\n\nNote: Credentials are never returned in the response for security."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateExchangeAccountRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New account name"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "New API key"
          },
          "api_secret": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Secret",
            "description": "New API secret"
          },
          "passphrase": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Passphrase",
            "description": "New passphrase"
          },
          "testnet": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Testnet",
            "description": "Whether to use testnet/sandbox"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "Whether the account is active"
          }
        },
        "type": "object",
        "title": "UpdateExchangeAccountRequest",
        "description": "Request to update an existing exchange account configuration."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}